requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
Pillow==10.1.0
PyMuPDF==1.23.8
pydantic==2.5.0
cors==1.0.1
redis==5.0.1
celery==5.3.4
websockets==12.0
google-cloud-storage==2.10.0
google-cloud-firestore>=2.11.0
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
anthropic>=0.37.0
transformers>=4.30.0
torch>=2.0.0
